Cache system lookups to save invocation time
Caching the output of the lookups performed drops time from 3-5
milliseconds to hundreds of nanoseconds, as calls to `os.environ` and
other system indicators are unlikely to change within the course of a
running process.

Using the `cache` decorator allows for passing in `no_color/force_color`
params to change the desired behavior, while allowing the text and
colors to be different without increasing the cache size.

Resolves #48

Signed-off-by: Mike Fiedler <miketheman@gmail.com>